Restore the functionality of the copy parameter of ucl_array_merge
We should operate on a reference (or copy) of elt, rather than elt directly.

Because kv_concat memcpy()s the pointer to the ucl_object_t, when you free elt, it results in a use-after-free when you try to read the values in top.
Doing a ucl_object_ref() will result in the items not being freeing until both elt and top are destroyed.

Optionally, using ucl_object_copy when the copy parameter of ucl_array_merge is true, will result in separate pointers, so future changes to top or elt do not affect the other.