Don't clobber the object's hash if it already exists
If we are in the 'merge' strategy, we replace the 'new' object with the 'old' object. This causes all keys in the old object to be erased, and replaced with an empty
hash, then the objects of the new key are added to that.

After this change, the original hash is preserved, and the keys of the new object are added to it.