Add missing include of <strings.h>.
libucl makes use of the strncasecmp() function, which according to POSIX
is provided by <strings.h>. On most operating systems, you can get away
with this, as it is also provided by <string.h>, but we shouldn't rely
on it.

This change is needed to get libucl to build on CloudABI, which has
standard header files that are far more picky, exactly so we can detect
these sorts of things.