Modernize CMake file with target-based includes.
This will associate include directories with individual targets
using target_include_directories() instead of setting them at the
directory level, which is the old style. This allows the library
to be used as a sub-directory of a parent project whose targets
depend on this one and CMake will handle transitive include
directories based on the PUBLIC/PRIVATE parameter values we give.

Also, create some namespace'd aliases for the two ucl targets,
which is also in line with modern style.

Tested building both the ucl and lua-ucl targets.

Fixes #190