Fix all Clang warnings from implicit conversions.
This commit will fix all remaining Clang warnings that arise
from building the core libucl library.  These warnings arise
from various kinds of conversions to and from char pointers:
{signed,unsigned} {const,non-const} char*.

The code uses all four of these variations and implicitly
converts between them in many places.  This generates warnings
on clang with -Wall.

To fix this without introducing too much noise in the code
there are four new macros:

  Cast to:               Macro
  --------------------------------
  char*                  S_(...)
  unsigned char*         U_(...)
  const char*            SC_(...)
  const unsigned char*   UC_(...)

The idea behind this change is, instead of changing any of
the types of e.g. function parameters or variables, we continue
to do the casting, but make it explicit with these macros.