Remove unused CMake logic and ad -Wno-pointer-sign.
There were a series of statements in the CMake file having
to do with controlling warning flags (including adding -Wall)
however they were not actually applied to the CMAKE_C_FLAGS
and so were unused.

This commit deletes that logic (which, if enabled, would
introduce many new warnings, especially the -std=c99).  But,
it leaves the -Wno-pointer-sign to disable that particular
warning which occurs in many places.