fix: Incorrect pointer arithmetics in ucl_expand_single_variable
This one I'm not 100% sure, but other code paths in that function are doing it this way.

As I said in previous issue, I had UclObjects holding bytes related to previous allocations. I *think* it's related to this code block. Valgrind showed read/write problems related to those pointers as well.

Hopefully this is last PR for this function.