ucl_maybe_parse_number: if there is trailing content, it is not a number
This prevents a string like `100GB_GPT_NTFS/100GB_GPT_NTFS.img` being
interpreted as `100GB`, and ending up with a value of `107374182400`

if you called `ucl_object_fromstring_common("100GB_GPT", 0, UCL_STRING_PARSE);`
it would match the cases for 'G' 'B' but not check `ucl_lex_is_atom_end()`
to ensure that what followed the 'B' was the end of an atom, and not more
string content.

Many of the other cases do, but this fixes the ones that do not.

Sponsored-by:	Klara Inc.