ucl_parse_value: fix out-of-bounds read
If the string ends with a Multiline terminator without a newline, the
function tried to read the next character to check for a newline without
checking if the pointer was past the end of the buffer. This commit adds
a bounds check and return early with an error in case of missing
newline.

Fix https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=21579