Fix including JSON files
When including a well-formed JSON file, the first opening brace is
ignored, resulting in a parser error when we find the corresponding
closing brace at the end of the document.

Track that the current stack of the parser state machine has seen an
opening brace when applicable, so that we do not choke when we find the
corresponding closing brace later.

Fixes #274