Fix regressions detected with new tests
When a file is included while reading a key, and this file starts with a
`{` char, a new stack element is not pushed and items are added to the
current one.  When we hit the corresponding `}` at the end of this file
we should therefore not pop the current stack element.