Allow Cmake to find Lua version between 5.4 and 5.1
Currently libUCL only searches for Lua 5.1 (when using LuaJIT)
and Lua 5.1 and 5.2. Commit introduces code that (if LuaJIT is
not used) seeking from Lua version 5.4 and if it not available
tries to seek version by version down to 5.1.
Seeking stops when Lua is found or if not it prints error
message to screen and stops configuring CMake build environment.